feat: centralize performance analytics state
